Move common logic to PNAdView
