+++ hybid.sdk/src/main/java/net/pubnative/lite/sdk/views/HyBidMRectAdView.java
@@ -32,9 +32,7 @@ import net.pubnative.lite.sdk.api.RequestManager;
 import net.pubnative.lite.sdk.banner.presenter.BannerPresenter;
 import net.pubnative.lite.sdk.mrect.presenter.MRectPresenterFactory;
 
-public class HyBidMRectAdView extends PNAdView implements BannerPresenter.Listener {
-
-    private BannerPresenter mPresenter;
+public class HyBidMRectAdView extends PNAdView {
 
     public HyBidMRectAdView(Context context) {
         super(context);
@@ -54,15 +52,6 @@ public class HyBidMRectAdView extends PNAdView implements BannerPresenter.Listen
     }
 
     @Override
-    protected void cleanup() {
-        super.cleanup();
-        if (mPresenter != null) {
-            mPresenter.destroy();
-            mPresenter = null;
-        }
-    }
-
-    @Override
     protected String getLogTag() {
         return HyBidMRectAdView.class.getSimpleName();
     }
@@ -73,48 +62,8 @@ public class HyBidMRectAdView extends PNAdView implements BannerPresenter.Listen
     }
 
     @Override
-    protected void renderAd() {
-        mPresenter = new MRectPresenterFactory(getContext())
+    protected BannerPresenter createPresenter() {
+        return new MRectPresenterFactory(getContext())
                 .createPresenter(mAd, this);
-        if (mPresenter != null) {
-            mPresenter.load();
-        } else {
-            invokeOnLoadFailed(new Exception("The server has returned an unsupported ad asset"));
-        }
-    }
-
-    @Override
-    protected void startTracking() {
-        if (mPresenter != null) {
-            mPresenter.startTracking();
-        }
-    }
-
-    @Override
-    protected void stopTracking() {
-        if (mPresenter != null) {
-            mPresenter.stopTracking();
-        }
-    }
-
-    //------------------------------ BannerPresenter Callbacks -------------------------------------
-
-    @Override
-    public void onBannerLoaded(BannerPresenter bannerPresenter, View banner) {
-        if (banner == null) {
-            invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
-        } else {
-            setupAdView(banner);
-        }
-    }
-
-    @Override
-    public void onBannerError(BannerPresenter bannerPresenter) {
-        invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
-    }
-
-    @Override
-    public void onBannerClicked(BannerPresenter bannerPresenter) {
-        invokeOnClick();
     }
 }
