+++ hybid.sdk/src/main/java/net/pubnative/lite/sdk/views/PNAdView.java
@@ -32,20 +32,25 @@ import android.view.ViewGroup;
 import android.widget.RelativeLayout;
 
 import net.pubnative.lite.sdk.api.RequestManager;
+import net.pubnative.lite.sdk.banner.presenter.BannerPresenter;
 import net.pubnative.lite.sdk.models.Ad;
 import net.pubnative.lite.sdk.utils.Logger;
 
-public abstract class PNAdView extends RelativeLayout implements RequestManager.RequestListener {
+public abstract class PNAdView extends RelativeLayout implements RequestManager.RequestListener, BannerPresenter.Listener {
 
     public interface Listener {
         void onAdLoaded();
+
         void onAdLoadFailed(Throwable error);
+
         void onAdImpression();
+
         void onAdClick();
     }
 
     private RequestManager mRequestManager;
     protected Listener mListener;
+    private BannerPresenter mPresenter;
     protected Ad mAd;
 
     public PNAdView(Context context) {
@@ -98,31 +103,38 @@ public abstract class PNAdView extends RelativeLayout implements RequestManager.
         setBackgroundColor(Color.TRANSPARENT);
         removeAllViews();
         mAd = null;
+
+        if (mPresenter != null) {
+            mPresenter.destroy();
+            mPresenter = null;
+        }
     }
 
     protected abstract String getLogTag();
 
     abstract RequestManager getRequestManager();
 
-    protected abstract void renderAd();
+    protected abstract BannerPresenter createPresenter();
 
-    protected abstract void startTracking();
-
-    protected abstract void stopTracking();
-
-    @Override
-    public void onRequestSuccess(Ad ad) {
-        if (ad == null) {
-            invokeOnLoadFailed(new Exception("Server returned null ad"));
+    protected void renderAd() {
+        mPresenter = createPresenter();
+        if (mPresenter != null) {
+            mPresenter.load();
         } else {
-            mAd = ad;
-            renderAd();
+            invokeOnLoadFailed(new Exception("The server has returned an unsupported ad asset"));
         }
     }
 
-    @Override
-    public void onRequestFail(Throwable throwable) {
-        invokeOnLoadFailed(new Exception(throwable));
+    protected void startTracking() {
+        if (mPresenter != null) {
+            mPresenter.startTracking();
+        }
+    }
+
+    protected void stopTracking() {
+        if (mPresenter != null) {
+            mPresenter.stopTracking();
+        }
     }
 
     protected void invokeOnLoadFinished() {
@@ -161,5 +173,41 @@ public abstract class PNAdView extends RelativeLayout implements RequestManager.
         invokeOnLoadFinished();
         startTracking();
         invokeOnImpression();
+    }
+
+    //----------------------------- BannerPresenter Callbacks --------------------------------------
+    @Override
+    public void onRequestSuccess(Ad ad) {
+        if (ad == null) {
+            invokeOnLoadFailed(new Exception("Server returned null ad"));
+        } else {
+            mAd = ad;
+            renderAd();
+        }
+    }
+
+    @Override
+    public void onRequestFail(Throwable throwable) {
+        invokeOnLoadFailed(new Exception(throwable));
+    }
+
+    //----------------------------- BannerPresenter Callbacks --------------------------------------
+    @Override
+    public void onBannerLoaded(BannerPresenter bannerPresenter, View banner) {
+        if (banner == null) {
+            invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
+        } else {
+            setupAdView(banner);
+        }
+    }
+
+    @Override
+    public void onBannerError(BannerPresenter bannerPresenter) {
+        invokeOnLoadFailed(new Exception("An error has occurred while rendering the ad"));
+    }
+
+    @Override
+    public void onBannerClicked(BannerPresenter bannerPresenter) {
+        invokeOnClick();
     }
 }
